requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
soundfile>=0.10.0
pyyaml>=6.0
tensorboard>=2.8.0
musdb>=0.4.0
museval>=0.4.0
einops>=0.6.0
tqdm>=4.62.0
